tt_project/requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
Django==3.0.8
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-filter==2.3.0
django-redis==4.12.1
django-timezone-field==4.0
djangorestframework==3.11.0
flower==0.9.5
gunicorn==20.0.4
humanize==2.5.0
kombu==4.6.11
Markdown==3.2.2
prometheus-client==0.8.0
psycopg2-binary==2.8.5
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
sentry-sdk==0.16.5
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.10
vine==1.3.0
